migration/requirements.txt
pymongo==4.7.3
psycopg2-binary==2.9.9
supabase==2.5.0
python-dotenv==1.0.1
tqdm==4.66.4
python-dateutil==2.9.0
